requirements.txt
Flask==3.1.0
flask-socketio==5.5.0
python-dotenv==1.0.0
openai==1.54.0
supabase==2.22.0
Flask-CORS==4.0.0